Multienterprise ERP

+++ connection_log.go
@@ -9,10 +9,11 @@ type ConnectionLog struct {
 	Id               int64      `json:"id"`
 	DateConnected    time.Time  `json:"dateConnected"`
 	DateDisconnected *time.Time `json:"dateDisconnected"`
-	User             int16      `json:"user"`
+	User             int32      `json:"user"`
 	Ok               bool       `json:"ok"`
 	IpAddress        string     `json:"ipAddress"`
 	UserName         string     `json:"userName"`
+	enterprise       int32
 }
 
 type ConnectionLogs struct {
@@ -22,8 +23,8 @@ type ConnectionLogs struct {
 
 func (q *PaginationQuery) getConnectionLogs() ConnectionLogs {
 	logs := make([]ConnectionLog, 0)
-	sqlStatement := `SELECT *,(SELECT username FROM "user" WHERE "user".id=connection_log."user") FROM public.connection_log ORDER BY date_connected DESC OFFSET $1 LIMIT $2`
-	rows, err := db.Query(sqlStatement, q.Offset, q.Limit)
+	sqlStatement := `SELECT *,(SELECT username FROM "user" WHERE "user".id=connection_log."user") FROM public.connection_log WHERE enterprise=$3 ORDER BY date_connected DESC OFFSET $1 LIMIT $2`
+	rows, err := db.Query(sqlStatement, q.Offset, q.Limit, q.Enterprise)
 	if err != nil {
 		log("DB", err.Error())
 		return ConnectionLogs{}
@@ -31,12 +32,12 @@ func (q *PaginationQuery) getConnectionLogs() ConnectionLogs {
 
 	for rows.Next() {
 		l := ConnectionLog{}
-		rows.Scan(&l.Id, &l.DateConnected, &l.DateDisconnected, &l.User, &l.Ok, &l.IpAddress, &l.UserName)
+		rows.Scan(&l.Id, &l.DateConnected, &l.DateDisconnected, &l.User, &l.Ok, &l.IpAddress, &l.enterprise, &l.UserName)
 		logs = append(logs, l)
 	}
 
-	sqlStatement = `SELECT COUNT(*) FROM public.connection_log`
-	row := db.QueryRow(sqlStatement)
+	sqlStatement = `SELECT COUNT(*) FROM public.connection_log WHERE enterprise=$1`
+	row := db.QueryRow(sqlStatement, q.Enterprise)
 	if row.Err() != nil {
 		log("DB", row.Err().Error())
 		return ConnectionLogs{}
@@ -49,15 +50,15 @@ func (q *PaginationQuery) getConnectionLogs() ConnectionLogs {
 }
 
 func (l *ConnectionLog) insertConnectionLog() {
-	sqlStatement := `INSERT INTO public.connection_log("user", ok, ip_address) VALUES ($1, $2, $3)`
-	db.Exec(sqlStatement, l.User, l.Ok, l.IpAddress)
+	sqlStatement := `INSERT INTO public.connection_log("user", ok, ip_address, enterprise) VALUES ($1, $2, $3, $4)`
+	db.Exec(sqlStatement, l.User, l.Ok, l.IpAddress, l.enterprise)
 }
 
 // Called during the client login.
 // 1. Logs the user connection
 // 2. Filters the user connection
-func userConnection(userId int16, ipAddress string) bool {
-	s := getSettingsRecord()
+func userConnection(userId int32, ipAddress string, enterpriseId int32) bool {
+	s := getSettingsRecordById(enterpriseId)
 	if !s.ConnectionLog {
 		return true
 	}
@@ -66,7 +67,7 @@ func userConnection(userId int16, ipAddress string) bool {
 	if strings.Contains(ipAddress, ":") {
 		ipAddress = ipAddress[:strings.Index(ipAddress, ":")]
 	}
-	l := ConnectionLog{User: userId, IpAddress: ipAddress}
+	l := ConnectionLog{User: userId, IpAddress: ipAddress, enterprise: enterpriseId}
 
 	// the default user ("marketnet") is not filtered
 	if userId == 1 {
@@ -102,24 +103,25 @@ func userConnection(userId int16, ipAddress string) bool {
 	return true
 }
 
-func userDisconnected(user int16) {
+func userDisconnected(user int32) {
 	sqlStatement := `UPDATE connection_log SET date_disconnected=CURRENT_TIMESTAMP(3) WHERE id=(SELECT id FROM public.connection_log WHERE "user"=$1 AND date_disconnected IS NULL ORDER BY date_connected DESC LIMIT 1)`
 	db.Exec(sqlStatement, user)
 }
 
 type ConnectionFilter struct {
-	Id        int16      `json:"id"`
-	Name      string     `json:"name"`
-	Type      string     `json:"type"` // I = IP, S = Schedule
-	IpAddress *string    `json:"ipAddress"`
-	TimeStart *time.Time `json:"timeStart"`
-	TimeEnd   *time.Time `json:"timeEnd"`
+	Id         int32      `json:"id"`
+	Name       string     `json:"name"`
+	Type       string     `json:"type"` // I = IP, S = Schedule
+	IpAddress  *string    `json:"ipAddress"`
+	TimeStart  *time.Time `json:"timeStart"`
+	TimeEnd    *time.Time `json:"timeEnd"`
+	enterprise int32
 }
 
-func getConnectionFilters() []ConnectionFilter {
+func getConnectionFilters(enterpriseId int32) []ConnectionFilter {
 	filters := make([]ConnectionFilter, 0)
-	sqlStatement := `SELECT * FROM public.connection_filter ORDER BY id ASC`
-	rows, err := db.Query(sqlStatement)
+	sqlStatement := `SELECT * FROM public.connection_filter WHERE enterprise=$1 ORDER BY id ASC`
+	rows, err := db.Query(sqlStatement, enterpriseId)
 	if err != nil {
 		log("DB", err.Error())
 		return filters
@@ -127,13 +129,13 @@ func getConnectionFilters() []ConnectionFilter {
 
 	for rows.Next() {
 		f := ConnectionFilter{}
-		rows.Scan(&f.Id, &f.Name, &f.Type, &f.IpAddress, &f.TimeStart, &f.TimeEnd)
+		rows.Scan(&f.Id, &f.Name, &f.Type, &f.IpAddress, &f.TimeStart, &f.TimeEnd, &f.enterprise)
 		filters = append(filters, f)
 	}
 	return filters
 }
 
-func getConnectionFilterRow(id int16) ConnectionFilter {
+func getConnectionFilterRow(id int32) ConnectionFilter {
 	f := ConnectionFilter{}
 	sqlStatement := `SELECT * FROM public.connection_filter WHERE id=$1 LIMIT 1`
 	row := db.QueryRow(sqlStatement, id)
@@ -142,12 +144,12 @@ func getConnectionFilterRow(id int16) ConnectionFilter {
 		return f
 	}
 
-	row.Scan(&f.Id, &f.Name, &f.Type, &f.IpAddress, &f.TimeStart, &f.TimeEnd)
+	row.Scan(&f.Id, &f.Name, &f.Type, &f.IpAddress, &f.TimeStart, &f.TimeEnd, &f.enterprise)
 
 	return f
 }
 
-func getConnectionFiltersByUser(userId int16) []ConnectionFilter {
+func getConnectionFiltersByUser(userId int32) []ConnectionFilter {
 	filters := make([]ConnectionFilter, 0)
 	sqlStatement := `SELECT connection_filter.* FROM public.connection_filter INNER JOIN connection_filter_user ON connection_filter_user.connection_filter=connection_filter.id WHERE connection_filter_user."user"=$1`
 	rows, err := db.Query(sqlStatement, userId)
@@ -158,7 +160,7 @@ func getConnectionFiltersByUser(userId int16) []ConnectionFilter {
 
 	for rows.Next() {
 		f := ConnectionFilter{}
-		rows.Scan(&f.Id, &f.Name, &f.Type, &f.IpAddress, &f.TimeStart, &f.TimeEnd)
+		rows.Scan(&f.Id, &f.Name, &f.Type, &f.IpAddress, &f.TimeStart, &f.TimeEnd, &f.enterprise)
 		filters = append(filters, f)
 	}
 	return filters
@@ -169,8 +171,8 @@ func (f *ConnectionFilter) insertConnectionFilter() bool {
 		return false
 	}
 
-	sqlStatement := `INSERT INTO public.connection_filter(name, type, ip_address, time_start, time_end) VALUES ($1, $2, $3, $4, $5)`
-	_, err := db.Exec(sqlStatement, f.Name, f.Type, f.IpAddress, f.TimeStart, f.TimeEnd)
+	sqlStatement := `INSERT INTO public.connection_filter(name, type, ip_address, time_start, time_end, enterprise) VALUES ($1, $2, $3, $4, $5, $6)`
+	_, err := db.Exec(sqlStatement, f.Name, f.Type, f.IpAddress, f.TimeStart, f.TimeEnd, f.enterprise)
 
 	if err != nil {
 		log("DB", err.Error())
@@ -191,8 +193,8 @@ func (f *ConnectionFilter) updateConnectionFilter() bool {
 		return false
 	}
 
-	sqlStatement := `UPDATE public.connection_filter SET name=$2, ip_address=$3, time_start=$4, time_end=$5 WHERE id=$1`
-	_, err := db.Exec(sqlStatement, f.Id, f.Name, f.IpAddress, f.TimeStart, f.TimeEnd)
+	sqlStatement := `UPDATE public.connection_filter SET name=$2, ip_address=$3, time_start=$4, time_end=$5 WHERE id=$1 AND enterprise=$6`
+	_, err := db.Exec(sqlStatement, f.Id, f.Name, f.IpAddress, f.TimeStart, f.TimeEnd, f.enterprise)
 
 	if err != nil {
 		log("DB", err.Error())
@@ -202,8 +204,8 @@ func (f *ConnectionFilter) updateConnectionFilter() bool {
 }
 
 func (f *ConnectionFilter) deleteConnectionFilter() bool {
-	sqlStatement := `DELETE FROM public.connection_filter WHERE id=$1`
-	_, err := db.Exec(sqlStatement, f.Id)
+	sqlStatement := `DELETE FROM public.connection_filter WHERE id=$1 AND enterprise=$2`
+	_, err := db.Exec(sqlStatement, f.Id, f.enterprise)
 
 	if err != nil {
 		log("DB", err.Error())
@@ -213,15 +215,15 @@ func (f *ConnectionFilter) deleteConnectionFilter() bool {
 }
 
 type ConnectionFilterUser struct {
-	ConnectionFilter int16  `json:"connectionFilter"`
-	User             int16  `json:"user"`
+	ConnectionFilter int32  `json:"connectionFilter"`
+	User             int32  `json:"user"`
 	UserName         string `json:"userName"`
 }
 
-func getConnectionFilterUser(filterId int16) []ConnectionFilterUser {
+func getConnectionFilterUser(filterId int32, enterpriseId int32) []ConnectionFilterUser {
 	filters := make([]ConnectionFilterUser, 0)
-	sqlStatement := `SELECT *,(SELECT username FROM "user" WHERE "user".id=connection_filter_user."user") FROM public.connection_filter_user WHERE connection_filter=$1 ORDER BY connection_filter ASC, "user" ASC`
-	rows, err := db.Query(sqlStatement, filterId)
+	sqlStatement := `SELECT *,(SELECT username FROM "user" WHERE "user".id=connection_filter_user."user") FROM public.connection_filter_user WHERE connection_filter=$1 AND (SELECT enterprise FROM connection_filter WHERE connection_filter.id=connection_filter_user.connection_filter)=$2 ORDER BY connection_filter ASC, "user" ASC`
+	rows, err := db.Query(sqlStatement, filterId, enterpriseId)
 	if err != nil {
 		log("DB", err.Error())
 		return filters
@@ -235,13 +237,31 @@ func getConnectionFilterUser(filterId int16) []ConnectionFilterUser {
 	return filters
 }
 
-func (f *ConnectionFilterUser) insertConnectionFilterUser() bool {
+func (f *ConnectionFilterUser) insertConnectionFilterUser(enterpriseId int32) bool {
+	filterInMemory := getConnectionFilterRow(f.ConnectionFilter)
+	if filterInMemory.enterprise != enterpriseId {
+		return false
+	}
+	userInMemory := getUserRow(f.User)
+	if userInMemory.enterprise != enterpriseId {
+		return false
+	}
+
 	sqlStatement := `INSERT INTO public.connection_filter_user(connection_filter, "user") VALUES ($1, $2)`
 	_, err := db.Exec(sqlStatement, f.ConnectionFilter, f.User)
 	return err == nil
 }
 
-func (f *ConnectionFilterUser) deleteConnectionFilterUser() bool {
+func (f *ConnectionFilterUser) deleteConnectionFilterUser(enterpriseId int32) bool {
+	filterInMemory := getConnectionFilterRow(f.ConnectionFilter)
+	if filterInMemory.enterprise != enterpriseId {
+		return false
+	}
+	userInMemory := getUserRow(f.User)
+	if userInMemory.enterprise != enterpriseId {
+		return false
+	}
+
 	sqlStatement := `DELETE FROM public.connection_filter_user WHERE connection_filter=$1 AND "user"=$2`
 	_, err := db.Exec(sqlStatement, f.ConnectionFilter, f.User)
 	return err == nil
